feat: Implement ServersProvider for managing server state and WebSocket integration
