+++ src/app/(gocontroller)/layout.tsx
@@ -3,6 +3,7 @@ import { SiteHeader } from "@/components/shell/site-header";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { auth } from "@/lib/auth";
 import { NotificationProvider } from "@/providers/notification-provider";
+import { ServersProvider } from "@/providers/servers-provider";
 import { routes } from "@/routes";
 import { redirect } from "next/navigation";
 
@@ -18,21 +19,23 @@ export default async function GoControllerLayout({
   }
 
   return (
-    <NotificationProvider>
-      <SidebarProvider
-        style={
-          {
-            "--sidebar-width": "calc(var(--spacing) * 72)",
-            "--header-height": "calc(var(--spacing) * 12)",
-          } as React.CSSProperties
-        }
-      >
-        <AppSidebar variant="inset" />
-        <SidebarInset>
-          <SiteHeader />
-          <div className="p-4 lg:p-6 h-full">{children}</div>
-        </SidebarInset>
-      </SidebarProvider>
-    </NotificationProvider>
+    <ServersProvider>
+      <NotificationProvider>
+        <SidebarProvider
+          style={
+            {
+              "--sidebar-width": "calc(var(--spacing) * 72)",
+              "--header-height": "calc(var(--spacing) * 12)",
+            } as React.CSSProperties
+          }
+        >
+          <AppSidebar variant="inset" />
+          <SidebarInset>
+            <SiteHeader />
+            <div className="p-4 lg:p-6 h-full">{children}</div>
+          </SidebarInset>
+        </SidebarProvider>
+      </NotificationProvider>
+    </ServersProvider>
   );
 }
